rename POTTools
