+++ sbncode/BeamSpillInfoRetriever/ICARUSBNBRetriever/ICARUSBNBRetriever_module.cc
@@ -12,7 +12,7 @@
 #include "art/Framework/Core/EDProducer.h"
 #include "messagefacility/MessageLogger/MessageLogger.h"
 #include "sbnobj/Common/POTAccounting/BNBSpillInfo.h"
-#include "sbncode/BeamSpillInfoRetriever/SBNDPOTTools.h"
+#include "sbncode/BeamSpillInfoRetriever/POTTools.h"
 #include "sbnobj/Common/Trigger/ExtraTriggerInfo.h"
 #include "sbndaq-artdaq-core/Overlays/ICARUS/ICARUSTriggerV3Fragment.hh"
 #include <sqlite3.h>
@@ -259,10 +259,6 @@ void sbn::ICARUSBNBRetriever::produce(art::Event& e)
   // We do not currently have the ability to figure out the first
   // spill that the DAQ was sensitive to, so don't try to save any
   // spill information
-  //
-  // TODO: long-term goal -- can we fix this?
-  // FIXME This is wrong.... 
-  //      Need to use:  ICARUSTriggerV3Fragment long getTotalTriggerBNBMaj() const
 
   if (e.event() == 1) return;
   
@@ -277,7 +273,7 @@ void sbn::ICARUSBNBRetriever::produce(art::Event& e)
   TotalBeamSpills += triggerInfo.number_of_gates_since_previous_event;
   MWRdata_t const MWRdata = extractSpillTimes(triggerInfo, bfp, bfp_mwr, fTimePad, MWRtoroidDelay, mwrdata);
   
-  int const spill_count = matchMultiWireData(e.id(), triggerInfo, MWRdata, e.event() == 1, fOutbeamInfos);
+  int const spill_count = matchMultiWireData(e.id(), triggerInfo, MWRdata, fOutbeamInfos);
   
   if(spill_count > int(triggerInfo.number_of_gates_since_previous_event))
     mf::LogDebug("ICARUSBNBRetriever")<< "Event Spills : " << spill_count << ", DAQ Spills : " << triggerInfo.number_of_gates_since_previous_event << " \t \t ::: WRONG!"<< std::endl;
@@ -357,37 +353,10 @@ int sbn::ICARUSBNBRetriever::matchMultiWireData(
   std::vector<int> matched_MWR;
   matched_MWR.resize(3);
   
-  
-  // NOTE: for now, this is dead code because we don't
-  // do anything for the first event in a run. We may want to revisit 
-  // this later to understand if there is a way we can do the POT
-  // accounting in the first event.
-  //
-  // Need to handle the first event in a run differently
-  if(isFirstEventInRun){
-    
-    //We'll remove the spills after our event
-    int spills_after_our_target = 0;
-    // iterate through all the spills to find the 
-    // spills that are after our triggered event
-    for (size_t i = 0; i < times_temps.size(); i++) {       
-      if(times_temps[i] > (triggerInfo.t_current_event+fTimePad)){
-	spills_after_our_target++;
-      }
-    }//end loop through spill times 	 
-    
-    // Remove the spills after our trigger
-    times_temps.erase(times_temps.end()-spills_after_our_target,times_temps.end());
-    
-    // Remove the spills before the start of our Run
-    times_temps.erase(times_temps.begin(), times_temps.end() - std::min(int(triggerInfo.number_of_gates_since_previous_event), int(times_temps.size())));
-        
-  }//end fix for "first event"
-  
-    ///reject time_stamps which have a trigger_type == 1 from data-base
-    //To-Do 
+  ///reject time_stamps which have a trigger_type == 1 from data-base
+  //To-Do 
 
-  //  mf::LogDebug("ICARUSBNBRetriever") << "Total number of Times we're going to test: " << times_temps.size() <<  std::endl;
+  // mf::LogDebug("ICARUSBNBRetriever") << "Total number of Times we're going to test: " << times_temps.size() <<  std::endl;
   // mf::LogDebug("ICARUSBNBRetriever") << std::setprecision(19) << "Upper Limit : " << (triggerInfo.t_current_event)+fTimePad <<  std::endl;
   // mf::LogDebug("ICARUSBNBRetriever") << std::setprecision(19) << "Lower Limit : " << (triggerInfo.t_previous_event)+fTimePad <<  std::endl;
