+++ sbncode/BeamSpillInfoRetriever/SBNDBNBEXTRetriever/SBNDBNBEXTRetriever_module.cc
@@ -8,7 +8,7 @@
 #include "art/Framework/Core/EDProducer.h"
 #include "messagefacility/MessageLogger/MessageLogger.h"
 #include "sbnobj/Common/POTAccounting/EXTCountInfo.h"
-#include "sbncode/BeamSpillInfoRetriever/SBNDPOTTools.h"
+#include "sbncode/BeamSpillInfoRetriever/POTTools.h"
 
 namespace sbn {
   class SBNDBNBEXTRetriever;
@@ -27,7 +27,6 @@ public:
   SBNDBNBEXTRetriever(SBNDBNBEXTRetriever &&) = delete;
   SBNDBNBEXTRetriever & operator = (SBNDBNBEXTRetriever const &) = delete;
   SBNDBNBEXTRetriever & operator = (SBNDBNBEXTRetriever &&) = delete;
-
 
 private:
   // Declare member data here.
