+++ sbncode/BeamSpillInfoRetriever/SBNDBNBZEROBIASRetriever/CMakeLists.txt
@@ -3,7 +3,7 @@ find_package(ifdh_art)
 
 cet_build_plugin(SBNDBNBZEROBIASRetriever art::module
     LIBRARIES
-        sbn_SBNDPOTTools
+        sbn_POTTools
 )
 
 install_headers()
